Change brace style to Allman
Allman seems to promote more modularity to me (since you can comment out the first line of a block and still works) and seems intuitive that the braces of a block should line up. Now that I have a full HD monitor, vertical space isn't at the premium that it once was either.